add small check in case we have race condition for moving duplicate workflows to GARDEN

+++ mpworks/workflows/wf_utils.py
@@ -75,5 +75,6 @@ def move_to_garden(m_dir):
     block_part = get_block_part(m_dir)
     garden_part = '/project/projectdirs/matgen/garden/'
     f_dir = os.path.join(garden_part, block_part)
-    shutil.move(m_dir, f_dir)
+    if os.path.exists(m_dir):
+        shutil.move(m_dir, f_dir)
     return f_dir
